pyproject.toml
[project]
name = "pmr-elirobots-driver"
version = "0.0.1"
description = ""
authors = [
    {name = "Felipe Gomes de Melo D'Elia",email = "felipegmelo@usp.br"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pmr_elirobots_msgs",
    "pyzmq",
    "tornado",
    "rich"
]

[tool.poetry]
packages = [{include = "pmr_elirobots_driver", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
